Split the special ops test cases
